重构 custom-form-data-table => custom-render 从easy-mock 对接更真实的api数据 解决选择图片后的报错
